tests cleanup after components extractions
